+++ host/class/uac/usb_host_uac/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this component will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- Changed API terminology from suspend/resume to pause/unpause as pre-requisite for root port suspend/resume feature
+
 ## [1.3.3] - 2025-11-27
 
 ### Changed